+++ skim/src/util.rs
@@ -136,7 +136,7 @@ impl LinePrinter {
         let w = ch.width().unwrap_or(2);
 
         if !skip {
-            let _ = canvas.put_cell(self.row, self.screen_col, style.apply(ch));
+            let _ = canvas.put_cell(self.row as u16, self.screen_col as u16, style.apply(ch));
         }
 
         self.screen_col += w;
@@ -282,9 +282,9 @@ pub fn reshape_string(
 /// 10% -> Size::Percent(10)
 pub fn margin_string_to_size(margin: &str) -> Size {
     if margin.ends_with('%') {
-        Size::Percent(min(100, margin[0..margin.len() - 1].parse::<usize>().unwrap_or(100)))
+        Size::Percent(min(100u16, margin[0..margin.len() - 1].parse::<u16>().unwrap_or(100)))
     } else {
-        Size::Fixed(margin.parse::<usize>().unwrap_or(0))
+        Size::Fixed(margin.parse::<u16>().unwrap_or(0))
     }
 }
